third_party/blink/web_tests/external/wpt/service-workers/service-worker/interfaces-sw.https-expected.txt

This is a testharness.js-based test.
PASS Interfaces and attributes in ServiceWorkerGlobalScope
FAIL test setup (cache creation) promise_test: Unhandled rejection with value: object "WebIDLParseError: Syntax error at line 1471, since `interface CustomElementRegistry`:
 define(DOMString name, CustomElementConstructor constructor, optional ElementDefinitionOptions
                        ^ Trailing comma in arguments list"
FAIL Event constructors assert_equals: FetchEvent.isReload should not exist expected (undefined) undefined but got (boolean) false
PASS xhr is not exposed
PASS URL.createObjectURL is not exposed
Harness: the test ran to completion.
